IntelligentBasedHMS/requirements.txt
streamlit==1.39.0
streamlit-lottie==0.0.5
tensorflow
scikit-learn==1.6.1
fastapi==0.115.0
Pillow==10.4.0
uvicorn==0.30.6
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
numpy==2.1.2
matplotlib==3.9.2
pathlib
sentence-transformers
langchain
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-community
langchain-experimental
langchain-huggingface
accelerate
transformers
torch
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint
langgraph-checkpoint-sqlite
psycopg2

